requirements.txt
# 多模态AI Agent - 网站部署版本
# 核心依赖
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0

# 网络请求
requests>=2.31.0

# 文件处理
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 工具
pydantic>=2.0.0
